extract class into a file
